make the fallback python implementation faster
I wanted to test the fallback python implementation to make sure
it still worked, but it was so slow that I couldn't even wait for
it to finish!  So I sped it up.

This fallback implementation should produce the same output as the
C version, although it will not have the same performance
characteristics.  (the code written here is O(n^2) for large systems
like the original code, my thought process being that the O(n^2) code
is cleaner, and that any system large enough to benefit from the
optimizations I added to the C code is already prohibitively large
for the python fallback to deal with!)