...and then rip it all out.
Keeping the legacy function around is an option (hence why I did it
initially), but unless we plan to add unit tests for it, it's not
a good idea to keep it around.

Tearing it out makes the code easier to maintain, and also makes
the diff against master smaller and easier to read.
(I also reintroduced some tabs where they originally appeared in
 the C file to make the diff even smaller)

This is a breaking change for anyone using the C library directly.
(...do people do that?)